Move models.py into app folder
