+++ database.py
@@ -11,5 +11,5 @@ Base.query = db_session.query_property()
 
 
 def init_db():
-    import models
+    from alayatodo import models
     Base.metadata.create_all(bind=engine)
